+++ LessonsAndersen/src/test/java/lesson16/task3/ParallelClass1.java
@@ -1,40 +1,34 @@
-package lesson16.task2;
-
+package lesson16.task3;
 import org.testng.annotations.Test;
 import static org.testng.AssertJUnit.assertTrue;
 
-public class ParallelClass2 {
+public class ParallelClass1 {
     @Test
-    public void prallel6() throws InterruptedException  {
+    public void prallel1() throws InterruptedException  {
         Thread.sleep(2000);
         assertTrue(true);
     }
     @Test
-    public void prallel7() throws InterruptedException  {
+    public void prallel2() throws InterruptedException  {
         Thread.sleep(2000);
         assertTrue(true);
     }
     @Test
-    public void prallel8() throws InterruptedException  {
-        Thread.sleep(2000);
-        assertTrue(true);
-    }
-
-    @Test
-    public void prallel9() throws InterruptedException  {
+    public void prallel3() throws InterruptedException  {
         Thread.sleep(2000);
         assertTrue(true);
     }
 
     @Test
-    public void prallel10() throws InterruptedException  {
+    public void prallel4() throws InterruptedException  {
+        Thread.sleep(2000);
+        assertTrue(true);
+    }
+
+    @Test
+    public void prallel5() throws InterruptedException  {
         Thread.sleep(2000);
         assertTrue(true);
     }
 }
-
-
-
-
-
-
+//task3
