+++ LessonsAndersen/src/test/java/lesson16/task4/DataProviderTest.java
@@ -4,12 +4,10 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 import utils.driver.consts.DriverSetUp;
 import java.time.Duration;
-
 public class DataProviderTest {
     @Test(dataProvider = "loginData")
     public void invalidLoginTest(String email, String passwd){
