+++ LessonsAndersen/src/test/resources/testngParametersHome.xml
@@ -1,6 +1,7 @@
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
 <suite name="LoginTestSuite" parallel="tests" thread-count="2">
     <test name="ChromeTest">
+<!--        task5-->
         <parameter name="browser" value="chrome"/>
         <classes>
             <class name="lesson16.task5.ParametersTest"/>
